src/main/java/au/com/unohomeloans/challenge/config/DynamoDBConfig.java
package au.com.unohomeloans.challenge.config;


import com.amazonaws.regions.Regions;
import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.annotation.Profile;

@Profile("prod")
@Configuration
public class DynamoDBConfig {

    @Bean
    public DynamoDBMapper dynamoDBMapperForProduction() {
        AmazonDynamoDB client = AmazonDynamoDBClientBuilder.standard()
                .withRegion(Regions.US_EAST_1)
                .build();
        return new DynamoDBMapper(client);
    }

}
